Refactor command routing logic codes
